Extract Redshift module into submodules
