+++ awswrangler/redshift/__init__.py
@@ -1,0 +1,17 @@
+"""Amazon Redshift Module."""
+
+from awswrangler.redshift._connect import connect, connect_temp
+from awswrangler.redshift._read import read_sql_query, read_sql_table, unload, unload_to_files
+from awswrangler.redshift._write import copy, copy_from_files, to_sql
+
+__all__ = [
+    "connect",
+    "connect_temp",
+    "copy",
+    "copy_from_files",
+    "read_sql_query",
+    "read_sql_table",
+    "to_sql",
+    "unload",
+    "unload_to_files",
+]
